requirements.txt
ansicon==1.89.0; platform_system == "Windows" and python_version >= "3.7"
blessed==1.19.1; python_version >= "3.7"
bpython==0.24; python_version >= "3.7"
certifi==2022.12.7; python_version >= "3.7" and python_version < "4"
charset-normalizer==3.0.1; python_version >= "3.7" and python_version < "4"
curtsies==0.4.1; python_version >= "3.7"
cwcwidth==0.1.8; python_version >= "3.7"
greenlet==2.0.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
icalendar==5.0.4; python_version >= "3.7"
idna==3.4; python_version >= "3.7" and python_version < "4"
jinxed==1.2.0; platform_system == "Windows" and python_version >= "3.7"
pygments==2.14.0; python_version >= "3.7"
python-dateutil==2.8.2; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7" \
    --hash=sha256:0123cacc1627ae19ddf3c27a5de5bd67ee4586fbdd6440d9748f8abb483d3e86 \
    --hash=sha256:961d03dc3453ebbc59dbdea9e4e11c5651520a876d0f4db161e8674aae935da9
pytz==2022.7.1; python_version >= "3.7"
pyxdg==0.28; python_version >= "3.7"
requests==2.28.2; python_version >= "3.7" and python_version < "4"
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7" \
    --hash=sha256:8abb2f1d86890a2dfb989f9a77cfcfd3e47c2a354b01111771326f8aa26e0254 \
    --hash=sha256:1e61c37477a1626458e36f7b1d82aa5c9b094fa4802892072e49de9c60c4c926
urllib3==1.26.14; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0"
wcwidth==0.2.6; python_version >= "3.7"
